Moved `pod_exec.go` into a separate package
additionally PodExec caller can now differentiate between command execution failures (exit code > 0) and other errors
